User management is as good as final for demo.

+++ grails-app/controllers/com/coveritas/heracles/ui/OrganizationController.groovy
@@ -28,33 +28,36 @@ class OrganizationController {
             notFound()
             return
         }
-
-        try {
-            Organization.withTransaction { status ->
-                // create org with mad and get UUID
-                Long userID = session['userID'] as Long
-                User user = User.get(userID)
-                Map<String,Object> result =  httpClientService.postParamsExpectMap('organization', [userUUID:user.uuid, userOrgUUID:user.organization.uuid])
-                String orgUUID = result.orgdUUID
-                String adminUUID = result.adminUUID
-                if (orgUUID) {
-                    Date now = new Date()
-                    organization.uuid = orgUUID
-                    organizationService.save(organization)
-                    User.create(adminUUID, "admin", organization, "@dm1n!", [Role.findByName(Role.ADMIN)] as Set<Role>)
+        Long userID = session['userID'] as Long
+        User u = User.get(userID)
+        if (u.isSysAdmin()) {
+            try {
+                Organization.withTransaction { status ->
+                    // create org with mad and get UUID
+                    Map<String, Object> result = httpClientService.postParamsExpectMap('organization', [userUUID: u.uuid, userOrgUUID: u.organization.uuid])
+                    String orgUUID = result.orgdUUID
+                    String adminUUID = result.adminUUID
+                    if (orgUUID) {
+                        Date now = new Date()
+                        organization.uuid = orgUUID
+                        organizationService.save(organization)
+                        User.create(adminUUID, "admin", organization, "@dm1n!", [Role.findByName(Role.ADMIN)] as Set<Role>)
+                    }
                 }
+            } catch (ValidationException e) {
+                respond organization.errors, view: 'create'
+                return
             }
-        } catch (ValidationException e) {
-            respond organization.errors, view:'create'
-            return
-        }
 
-        request.withFormat {
-            form multipartForm {
-                flash.message = message(code: 'default.created.message', args: [message(code: 'organization.label', default: 'Organization'), organization.id])
-                redirect organization
+            request.withFormat {
+                form multipartForm {
+                    flash.message = message(code: 'default.created.message', args: [message(code: 'organization.label', default: 'Organization'), organization.id])
+                    redirect organization
+                }
+                '*' { respond organization, [status: CREATED] }
             }
-            '*' { respond organization, [status: CREATED] }
+        } else {
+            notAllowed('default.not.created.message')
         }
     }
 
@@ -67,20 +70,24 @@ class OrganizationController {
             notFound()
             return
         }
-
-        try {
-            organizationService.save(organization)
-        } catch (ValidationException e) {
-            respond organization.errors, view:'edit'
-            return
-        }
-
-        request.withFormat {
-            form multipartForm {
-                flash.message = message(code: 'default.updated.message', args: [message(code: 'organization.label', default: 'Organization'), organization.name])
-                redirect organization
+        Long userID = session['userID'] as Long
+        User u = User.get(userID)
+        if ( u.isAdmin(organization)||u.isSysAdmin()) {
+            try {
+                organizationService.save(organization)
+            } catch (ValidationException e) {
+                respond organization.errors, view: 'edit'
+                return
             }
-            '*'{ respond organization, [status: OK] }
+            request.withFormat {
+                form multipartForm {
+                    flash.message = message(code: 'default.updated.message', args: [message(code: 'organization.label', default: 'Organization'), organization.name])
+                    redirect organization
+                }
+                '*' { respond organization, [status: OK] }
+            }
+        } else {
+            notAllowed('default.not.updated.message')
         }
     }
 
@@ -89,25 +96,40 @@ class OrganizationController {
             notFound()
             return
         }
+        Long userID = session['userID'] as Long
+        User u = User.get(userID)
+        if ( u.isAdmin(organization)||u.isSysAdmin()) {
+            organizationService.delete(id)
 
-        organizationService.delete(id)
-
-        request.withFormat {
-            form multipartForm {
-                flash.message = message(code: 'default.deleted.message', args: [message(code: 'organization.label', default: 'Organization'), id])
-                redirect action:"index", method:"GET"
+            request.withFormat {
+                form multipartForm {
+                    flash.message = message(code: 'default.deleted.message', args: [message(code: 'organization.label', default: 'Organization'), id])
+                    redirect action: "index", method: "GET"
+                }
+                '*' { render status: NO_CONTENT }
             }
-            '*'{ render status: NO_CONTENT }
+        } else {
+            notAllowed('default.not.deleted.message')
         }
     }
 
     protected void notFound() {
         request.withFormat {
             form multipartForm {
-                flash.message = message(code: 'default.not.found.message', args: [message(code: 'organization.label', default: 'Organization'), params.id])
+                flash.message = message(code: 'default.not.found.message', args: [message(code: 'organization.label', default: 'Organization'), params.name])
                 redirect action: "index", method: "GET"
             }
             '*'{ render status: NOT_FOUND }
+        }
+    }
+
+    protected void notAllowed(String msg) {
+        request.withFormat {
+            form multipartForm {
+                flash.message = message(code: msg, args: [message(code: 'organization.label', default: 'Organization'), params.name])
+                redirect action: "index", method: "GET"
+            }
+            '*'{ render status: FORBIDDEN }
         }
     }
 }
